[fix]>>等を単一のトークンとして扱うよう修正
