+++ srcs/tokenizer/validate_line.c
@@ -1,93 +1,97 @@
 #include "minishell.h"
 
-/*
-** line中にトークンの不正な連続がないか確認する。
-** ;; || & ";[space];" "|[space]|" ";|"
-*/
-
 #define AMP		'&'
 #define OPS		"|;&<>"
 #define SEP		"|;&"
 #define SPACES	"\v\r\f\t\n "
 
-static size_t	count_consective_char(char *line)
+static char	*skip_inquote(char *line, char quote)
 {
-	size_t	i;
-	char	head;
-
-	i = 0;
-	head = *line;
-	while (*line && line[i] == head)
-		i++;
-	return (i);
+	line++;
+	while (*line != quote)
+		line++;
+	return (line);
 }
 
-static bool	error_return(char *line, char *prev_ptr)
+static bool	check_redirect(char *line)
 {
-	const char	*newline = "newline";
-	size_t		i;
-
-	ft_putstr_fd("minishell: syntax error near unexpected token `", 2);
-	i = count_consective_char(prev_ptr);
-	if (ft_strchr(SEP, *prev_ptr) && i >= 2)
-		i = 2;
-	if (*prev_ptr == '<' && line != prev_ptr)
-		i = i % 2 + 1;
-	if (*prev_ptr == '>' && line != prev_ptr)
-		i = i % 3 + 1;
-	if (ft_strchr("<>", *prev_ptr) && !*line)
-		ft_putstr_fd((char *)newline, STDERR_FILENO);
-	else
-		while (i--)
-			ft_putchar_fd(*prev_ptr, STDERR_FILENO);
-	ft_putendl_fd("'", STDERR_FILENO);
-	store_exitstatus(SAVE, 258);
-	return (false);
-}
-
-static bool	check_operator(char *line)
-{
-	char	*prev_ptr;
-
 	while (*line != '\0')
 	{
-		prev_ptr = line;
+		if (!ft_strchr(OPS, *line) && !ft_strchr(SPACES, *line))
+			return (true);
+		if (ft_strchr(SEP, *line))
+			return (error_return(line, 'a', false));
+		if (ft_strchr("<>", *line))
+		{
+			if (*(line + 1) == *line)
+				return (error_return(line, *line, false));
+			return (error_return(line, 'a', false));
+		}
 		line++;
+	}
+	return (error_return(line, 'a', false));
+}
+
+
+static bool	check_operator(char *line, char last_op)
+{
+	bool	has_space;
+
+	has_space = false;
+	line++;
+	while (*line != '\0')
+	{
 		if (!ft_strchr(OPS, *line) && !ft_strchr(SPACES, *line))
 			return (true);
 		if (ft_strchr(SPACES, *line))
-		{
-			line++;
-			if (ft_strchr("<>", *prev_ptr) && ft_strchr(OPS, *line))
-				return (error_return(line, line));
-			continue ;
-		}
-		if (*prev_ptr == '>' && count_consective_char(prev_ptr) > 2)
-			return (error_return(line, prev_ptr));
-		if (ft_strchr("|;&<", *line) && *line != *prev_ptr)
-			prev_ptr = line;
-		if (ft_strchr("|;&<", *line))
-			return (error_return(line, prev_ptr));
+			has_space = true;
+		else if (ft_strchr("<>", *line) && last_op == *line && !has_space)
+			return (check_redirect(line + 1));
+		else if (ft_strchr("<>", *line))
+			return (error_return(line, last_op, has_space));
+		else if (ft_strchr(SEP, *line))
+			return (error_return(line, last_op, has_space));
+		line++;
 	}
-	if (*prev_ptr == ';')
+	if (last_op == ';')
 		return (true);
-	return (error_return(line, prev_ptr));
+	else
+		return (error_return(line, last_op, false));
+}
+
+static bool	check_first_segment(char *line)
+{
+	char	*first_sep;
+	char	sep;
+
+	first_sep = get_first_sep(line);
+	if (first_sep == NULL)
+		return (true);
+	sep = *first_sep;
+	while (*line != sep)
+	{
+		if (!ft_strchr(OPS, *line) && !ft_strchr(SPACES, *line))
+			return (true);
+		line++;
+	}
+	if (*(first_sep + 1) == sep)
+		return (error_return(first_sep, sep, false));
+	return (error_return(first_sep, 'a', false));
 }
 
 bool	validate_line(char *line)
 {
-	char	*head;
-
-	head = line;
+	if (!check_first_segment(line))
+		return (false);
 	while (*line != '\0')
 	{
+		if (*line == '"' || *line == '\'')
+			line = skip_inquote(line, *line);
 		if (ft_strchr(OPS, *line))
 		{
-			if (line == head && !ft_strchr("<>", *line))
-				return (error_return(line, line));
 			if (*line == AMP)
-				return (error_return(line, line));
-			if (!check_operator(line))
+				return (error_return(line, 'n', false));
+			if (!check_operator(line, *line))
 				return (false);
 		}
 		line++;
